+++ pkg/ring/lifecycler.go
@@ -27,6 +27,17 @@ var (
 	errInvalidTokensGeneratorStrategy = errors.New("invalid token generator strategy")
 )
 
+type LifecyclerDelegate interface {
+	// OnRingInstanceHeartbeat is called while the instance is updating its heartbeat
+	// in the ring.
+	OnRingInstanceHeartbeat(lifecycler *Lifecycler, ringDesc *Desc)
+}
+
+type DefaultLifecyclerDelegate struct{}
+
+func (d DefaultLifecyclerDelegate) OnRingInstanceHeartbeat(lifecycler *Lifecycler, ringDesc *Desc) {
+}
+
 // LifecyclerConfig is the config to build a Lifecycler.
 type LifecyclerConfig struct {
 	RingConfig Config `yaml:"ring"`
@@ -108,6 +119,7 @@ type Lifecycler struct {
 	cfg             LifecyclerConfig
 	flushTransferer FlushTransferer
 	KVStore         kv.Client
+	delegate        LifecyclerDelegate
 
 	actorChan    chan func()
 	autojoinChan chan struct{}
@@ -148,6 +160,22 @@ type Lifecycler struct {
 	logger            log.Logger
 
 	tg TokenGenerator
+}
+
+func NewLifecyclerWithDelegate(
+	cfg LifecyclerConfig,
+	flushTransferer FlushTransferer,
+	ringName, ringKey string,
+	autoJoinOnStartup, flushOnShutdown bool,
+	logger log.Logger,
+	reg prometheus.Registerer,
+	delegate LifecyclerDelegate,
+) (*Lifecycler, error) {
+	l, err := NewLifecycler(cfg, flushTransferer, ringName, ringKey, autoJoinOnStartup, flushOnShutdown, logger, reg)
+	if l != nil {
+		l.delegate = delegate
+	}
+	return l, err
 }
 
 // NewLifecycler creates new Lifecycler. It must be started via StartAsync.
@@ -209,6 +237,7 @@ func NewLifecycler(
 		lifecyclerMetrics:    NewLifecyclerMetrics(ringName, reg),
 		logger:               logger,
 		tg:                   tg,
+		delegate:             &DefaultLifecyclerDelegate{},
 	}
 
 	l.lifecyclerMetrics.tokensToOwn.Set(float64(cfg.NumTokens))
@@ -973,6 +1002,7 @@ func (i *Lifecycler) updateConsul(ctx context.Context) error {
 			instanceDesc.RegisteredTimestamp = i.getRegisteredAt().Unix()
 			ringDesc.Ingesters[i.ID] = instanceDesc
 		}
+		i.delegate.OnRingInstanceHeartbeat(i, ringDesc)
 
 		return ringDesc, true, nil
 	})